aeon/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
deprecation==2.1.0
Django==4.2.16
django-colorfield==0.11.0
django-email-verification==0.3.3
idna==3.10
mysqlclient==2.2.4
packaging==24.1
pillow==10.4.0
PyJWT==2.9.0
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
validators==0.34.0
